refactor(rasterize-list): Rename to more descriptive write image
