+++ src/utils/write-images.js
